Estimate via Total least square approach for two stage bootstrap

Estimate via Total least square approach for two stage bootstrap

tlsLm.boot(X, Y, nruns.X, B = 100)

Arguments

X

n*k matrix, including k predictors

Y

n*1 matrix, the observations

nruns.X

the number of runs used for computing each columns of X

Value

a list containing the estimate and confidence interval of the scaling factors estimate.

coefficient: a k vector, the scaling factors best-estimates

confidence interval: the lower and upper bounds of the confidence interval on each scaling factor.

dcons: the variable used in the residual consistency check.

X.tilde: a k*n matrix, the reconstructed responses patterns TLS fit,

Y.tilde: a 1*n matrix, the reconstructed observations TLS fit.